requirements.txt
langchain>=0.3.25
langchain-openai>=0.3.16
langchain-community>=0.3.23
python-dotenv>=1.0.1
openai>=1.77.0
# crewai>=0.28.0
langsmith>=0.3.42